chore(emacs): remove configuration
I really wanted to make it work, because .org is amazing, and the
plugin system is fairly more advanced, but the truth is, I can't get
used to how slow it feels, and how brittle package interactions can be.

I did love the fact that you could do litterate programming, and some of
the niceties around prose writing, but that wasn't enough for me.