feat(yabai): adopt it for real
It's a bit buggy, which is a shame, but ultimately, it's the only decent
macOS WM. We get Stage Manager, but not window snapping or proper
tiling. Still better than realising the WM that made you feel FOMO is
built by a bunch of unsavoury people, but that's a low bar to clear
Apple.