feat(nvim): move lsp stuff to lazy plugin spec
Wow, this was awful to migrate. Finding a way that wasn't awkward and
didn't break nvim-cmp silently was a lot harder than I would've liked.
It was so bad that I shopped around for another plugin. I haven't made
the switch yet, because that sounds like a pain, but at the next whiff
of nvim-cmp making my life harder, it's out.

I still find it mind-boggling that neovim understands the lsp protocol,
has a completion engine, but requires you to use a plugin just to glue
the two of them. It's seemingly going to get easier as there's been some
discussions in the official repo around October, but it's not guaranteed
(which is fair for any FOSS project).