feat(nvim): use nightly
neovim doesn't really release features outside of new minors (it's still
pre 1.0), so they tend to stack features on the nightly to make a big
release.

It wasn't such a bad thing when they released minors more often, but
we're at almost a year without a new one this time around. Looking
forward to not needing that overlay, because it didn't feel good to add
so many inputs.